* warnings-- for several other plugins (signed/unsigned issues)
git-svn-id: file:///.data/svn_convert/svn@8319 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-playground/plugins/netmonitor/monitor.cpp
@@ -262,8 +262,8 @@ void MonitorWindow::outputLog()
 
     QMutexLocker lock(&m_mutex);
 
-    for(unsigned i = 0; i < m_logStrings.count(); i++)
-        edit->append(m_logStrings[(int)i]);
+    for(int i = 0; i < m_logStrings.count(); i++)
+        edit->append(m_logStrings[i]);
 
     m_logStrings.clear();
     if (bAutoscroll)
